Created AbstractEntity, User, and UserRepository classes
