+++ src/main/java/org/launchcode/liftoffgroupproject/models/UserRepository.java
@@ -1,0 +1,9 @@
+package org.launchcode.liftoffgroupproject.models;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface UserRepository extends CrudRepository<User, Integer>{
+
+    User findByUsername(String username);
+
+}
